+++ src/plugin-info.ts
@@ -1,5 +1,5 @@
 /* File will be overwritten by bin/release.sh! */
 export const PLUGIN_INFO = {
-  "pluginVersion": "1.6.4",
-  "pluginReleasedAt": "2024-09-20T16:30:33+0200"
+  "pluginVersion": "1.6.5",
+  "pluginReleasedAt": "2024-10-09T13:11:11+0200"
 }