+++ src/routes/note.ts
@@ -49,6 +49,7 @@ import {
 } from "src/utils/parameters";
 import {
   checkForEnabledPeriodicNoteFeature,
+  createPeriodicNote,
   getAllPeriodicNotes,
   PeriodicNoteType,
 } from "src/utils/periodic-notes-handling";
@@ -58,7 +59,8 @@ import {
   escapeRegExpChars,
   parseStringIntoRegex,
 } from "src/utils/string-handling";
-import { focusOrOpenFile } from "src/utils/ui";
+
+import { focusOrOpenNote } from "src/utils/ui";
 import { zodOptionalBoolean, zodSanitizedNotePath } from "src/utils/zod";
 
 // SCHEMATA ----------------------------------------
@@ -273,7 +275,7 @@ async function handleGet(
 ): Promise<HandlerFileSuccess | HandlerFailure> {
   const { _resolved: { inputPath }, silent } = params;
   const res = await getNoteDetails(inputPath);
-  if (res.isSuccess && !silent) await focusOrOpenFile(inputPath);
+  if (res.isSuccess && !silent) await focusOrOpenNote(inputPath);
   return res;
 }
 
@@ -383,6 +385,7 @@ async function handleAppend(
     ["create-if-not-found"]: shouldCreateNote,
     ["ensure-newline"]: shouldEnsureNewline,
     ["if-headline-missing"]: ifHeadlineMissing,
+    ["periodic-note"]: periodicNoteType,
     content,
     silent,
     uid,
@@ -422,24 +425,38 @@ async function handleAppend(
       return failure(ErrorCode.NotFound, STRINGS.note_not_found);
     }
 
-    // We're supposed to create the note. We try to create it.
-    const resCreate = await createNote(path, "");
-    if (!resCreate.isSuccess) return resCreate;
+    // We're supposed to create the note!
 
-    // If the note was requested via UID, we need to set the UID in the front
-    // matter of the newly created note.
-    if (inputKey === NoteTargetingParameterKey.UID) {
-      await this.app.fileManager.processFrontMatter(
-        resCreate.result,
-        (fm) => fm[this.settings.frontmatterKey] = uid!,
-      );
+    // The requested note is a periodic note.
+    if (inputKey === NoteTargetingParameterKey.PeriodicNote) {
+      const newNote = await createPeriodicNote(periodicNoteType!);
+      if (!newNote) {
+        return failure(
+          ErrorCode.UnableToCreateNote,
+          STRINGS.unable_to_write_note,
+        );
+      }
+    } //
+    // The requested note is not a periodic note.
+    else {
+      const resCreate = await createNote(path, "");
+      if (!resCreate.isSuccess) return resCreate;
+
+      // If the note was requested via UID, we need to set the UID in the front
+      // matter of the newly created note.
+      if (inputKey === NoteTargetingParameterKey.UID) {
+        await this.app.fileManager.processFrontMatter(
+          resCreate.result,
+          (fm) => fm[this.settings.frontmatterKey] = uid!,
+        );
+      }
     }
   }
 
   // Manipulate the file.
   const resAppend = await appendAsRequested();
   if (resAppend.isSuccess) {
-    if (!silent) await focusOrOpenFile(path);
+    if (!silent) await focusOrOpenNote(path);
     return success({ message: resAppend.result }, path);
   }
   return resAppend;
@@ -456,6 +473,7 @@ async function handlePrepend(
     ["ensure-newline"]: shouldEnsureNewline,
     ["ignore-front-matter"]: shouldIgnoreFrontMatter,
     ["if-headline-missing"]: ifHeadlineMissing,
+    ["periodic-note"]: periodicNoteType,
     content,
     silent,
     uid,
@@ -508,24 +526,38 @@ async function handlePrepend(
       return failure(ErrorCode.NotFound, STRINGS.note_not_found);
     }
 
-    // We're supposed to create the note. We try to create it.
-    const resCreate = await createNote(path, "");
-    if (!resCreate.isSuccess) return resCreate;
+    // We're supposed to create the note!
 
-    // If the note was requested via UID, we need to set the UID in the front
-    // matter of the newly created note.
-    if (inputKey === NoteTargetingParameterKey.UID) {
-      await this.app.fileManager.processFrontMatter(
-        resCreate.result,
-        (fm) => fm[this.settings.frontmatterKey] = uid!,
-      );
+    // The requested note is a periodic note.
+    if (inputKey === NoteTargetingParameterKey.PeriodicNote) {
+      const newNote = await createPeriodicNote(periodicNoteType!);
+      if (!newNote) {
+        return failure(
+          ErrorCode.UnableToCreateNote,
+          STRINGS.unable_to_write_note,
+        );
+      }
+    } //
+    // The requested note is not a periodic note.
+    else {
+      const resCreate = await createNote(path, "");
+      if (!resCreate.isSuccess) return resCreate;
+
+      // If the note was requested via UID, we need to set the UID in the front
+      // matter of the newly created note.
+      if (inputKey === NoteTargetingParameterKey.UID) {
+        await this.app.fileManager.processFrontMatter(
+          resCreate.result,
+          (fm) => fm[this.settings.frontmatterKey] = uid!,
+        );
+      }
     }
   }
 
   // Manipulate the file.
   const resPrepend = await prependAsRequested();
   if (resPrepend.isSuccess) {
-    if (!silent) await focusOrOpenFile(path);
+    if (!silent) await focusOrOpenNote(path);
     return success({ message: resPrepend.result }, path);
   }
   return resPrepend;
@@ -538,7 +570,7 @@ async function handleTouch(
 
   const res = await touchNote(inputPath);
   if (!res.isSuccess) return res;
-  if (!silent) await focusOrOpenFile(inputPath);
+  if (!silent) await focusOrOpenNote(inputPath);
   return success({ message: STRINGS.touch_done }, inputPath);
 }
 
@@ -549,7 +581,7 @@ async function handleSearchStringAndReplace(
 
   const res = await searchAndReplaceInNote(inputPath, search, replace);
   if (!res.isSuccess) return res;
-  if (!silent) await focusOrOpenFile(inputPath);
+  if (!silent) await focusOrOpenNote(inputPath);
   return success({ message: res.result }, inputPath);
 }
 
@@ -563,7 +595,7 @@ async function handleSearchRegexAndReplace(
 
   const res = await searchAndReplaceInNote(inputPath, resSir.result, replace);
   if (!res.isSuccess) return res;
-  if (!silent) await focusOrOpenFile(inputPath);
+  if (!silent) await focusOrOpenNote(inputPath);
   return success({ message: res.result }, inputPath);
 }